requirements.txt
pandas
pyyaml
snowflake-connector-python
yq
